requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
blis==0.7.4
cachetools==4.2.2
catalogue==2.0.4
certifi==2021.5.30
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cymem==2.0.5
discord.py==1.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
fuzzywuzzy==0.18.0
google-api-core==1.30.0
google-api-python-client==1.12.8
google-auth==1.32.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httplib2==0.19.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
Jinja2==3.0.1
langdetect==1.0.9
MarkupSafe==2.0.1
multidict==5.1.0
murmurhash==1.0.5
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pathy==0.5.2
Pillow==8.3.0
preshed==3.0.5
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.7.4
pyfiglet==0.8.post1
pyparsing==2.4.7
python-dateutil==2.8.1
python-Levenshtein==0.12.2
python-resize-image==1.1.19
pytz==2021.1
requests==2.25.1
rfc3986==1.5.0
rsa==4.7.2
six==1.16.0
smart-open==3.0.0
sniffio==1.2.0
soupsieve==2.2.1
spacy==3.0.6
spacy-legacy==3.0.5
srsly==2.4.1
tabulate==0.8.9
termcolor==1.1.0
thinc==8.0.3
tqdm==4.61.0
typer==0.3.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
wasabi==0.8.2
wikipedia==1.4.0
yarl==1.6.3
